# docs/swagger.yaml
basePath: /
definitions:
  models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  server.webhookMessage:
    properties:
      content:
        type: string
      receivers:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - title
    type: object
  service.ComponentStatus:
    properties:
      latency_ms:
        type: number
      message:
        type: string
      status:
        type: string
    type: object
  service.DBPoolStats:
    properties:
      idle:
        type: integer
      in_use:
        type: integer
      max_idle_closed:
        type: integer
      max_lifetime_closed:
        type: integer
      max_open:
        type: integer
      open_connections:
        type: integer
      wait_count:
        type: integer
      wait_duration:
        type: string
    type: object
  service.HealthReport:
    properties:
      components:
        additionalProperties:
          $ref: '#/definitions/service.ComponentStatus'
        type: object
      db_pool:
        $ref: '#/definitions/service.DBPoolStats'
      go_version:
        type: string
      goroutines:
        type: integer
      memory:
        $ref: '#/definitions/service.MemStats'
      status:
        type: string
      uptime:
        type: string
    type: object
  service.MemStats:
    properties:
      alloc_mb:
        type: number
      num_gc:
        type: integer
      sys_mb:
        type: number
    type: object
info:
  contact:
    name: Eric Jiang
    url: http://www.swagger.io/support
  description: 'Development Environment: go v1.23.7 + gin v1.10.1 + gorm v1.30.2 +
    viper v1.20.1'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: 情感沙滩 API
  version: "1.0"
paths:
  /callback:
    get:
      responses: {}
      summary: GitHub 回调接口
      tags:
      - 注册登陆
  /health:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 服务健康
          schema:
            $ref: '#/definitions/service.HealthReport'
        "503":
          description: 服务降级
          schema:
            $ref: '#/definitions/service.HealthReport'
      summary: 健康检查
      tags:
      - 系统
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录参数
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses: {}
      summary: 登陆获取 Token
      tags:
      - 注册登陆
  /login/github:
    get:
      responses: {}
      summary: GitHub 登录
      tags:
      - 注册登陆
  /register:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Name
        in: formData
        name: name
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      - description: repeat_password
        in: formData
        name: repeat_password
        required: true
        type: string
      - description: Phone
        in: formData
        name: phone
        required: true
        type: string
      - description: EMAIL
        in: formData
        name: email
        type: string
      produces:
      - application/json
      responses: {}
      summary: 创建用户
      tags:
      - 注册登陆
  /v1/api/webhook:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求参数
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/server.webhookMessage'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Webhook 对外接口
      tags:
      - API
  /v1/relation/add:
    post:
      parameters:
      - description: 发起方用户 ID
        in: formData
        name: userId
        type: integer
      - description: 目标昵称或目标用户 ID
        in: formData
        name: targetName
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 通过昵称加好友
      tags:
      - 好友关系
  /v1/relation/list:
    post:
      parameters:
      - description: 用户 ID
        in: formData
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 获取好友列表
      tags:
      - 好友关系
  /v1/user/condition:
    get:
      parameters:
      - description: 用户 ID
        in: query
        name: id
        type: string
      - description: 手机号
        in: query
        name: phone
        type: string
      - description: Email
        in: query
        name: email
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 按条件查询用户
      tags:
      - 用户
  /v1/user/delete:
    delete:
      parameters:
      - description: ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 删除用户
      tags:
      - 用户
  /v1/user/list:
    get:
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 获取所有用户
      tags:
      - 用户
  /v1/user/update:
    post:
      parameters:
      - description: ID
        in: formData
        name: id
        required: true
        type: string
      - description: 用户名
        in: formData
        name: name
        type: string
      - description: 密码
        in: formData
        name: password
        type: string
      - description: 手机号
        in: formData
        name: phone
        type: string
      - description: Email
        in: formData
        name: email
        type: string
      - description: 头像
        in: formData
        name: avatar
        type: string
      - description: 性别
        in: formData
        name: gender
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 更新用户信息
      tags:
      - 用户
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"